refactor visualizations
