+++ client/src/model/VisualizationModels/ScatterPlotMatrixModel.js
@@ -1,0 +1,9 @@
+import VisualizationModel from '../VisualizationModel';
+
+class ScatterPlotMatrixModel extends VisualizationModel {
+  getPreparedDataset() {
+    return this.dataModel.getSelectedDataset();
+  }
+}
+
+export default ScatterPlotMatrixModel;
